refactor(runner): split test fixtures by responsibility
