+++ crates/runner/src/cmd/gc/images/tests.rs
@@ -8,7 +8,9 @@ use crate::cmd::gc::test_support::{
     assert_is_symlink, old_gc_time, set_mtime, set_soft_nofile_limit_for_child, test_home,
 };
 use crate::lock;
-use crate::test_fixtures::{ignored_child_test_env_guard_enabled, run_ignored_child_test};
+use crate::test_fixtures::ignored_child::{
+    ignored_child_test_env_guard_enabled, run_ignored_child_test,
+};
 
 fn rootfs_state_with_deletion(
     rootfs_path: PathBuf,
