+++ crates/runner/src/host_file.rs
@@ -605,7 +605,9 @@ mod tests {
     use std::time::Duration;
 
     use super::*;
-    use crate::test_fixtures::{ignored_child_test_env_guard_enabled, run_ignored_child_test};
+    use crate::test_fixtures::ignored_child::{
+        ignored_child_test_env_guard_enabled, run_ignored_child_test,
+    };
 
     const RESTRICTIVE_UMASK_CHILD_ENV: &str = "VM0_RUN_RESTRICTIVE_UMASK_TEST";
