+++ crates/runner/src/provider/mod.rs
@@ -541,7 +541,7 @@ pub trait JobProvider: Send + Sync {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_fixtures::execution_context_for_test;
+    use crate::test_fixtures::execution_context::execution_context_for_test;
 
     fn minimal_context(run_id: RunId) -> ExecutionContext {
         let mut ctx = execution_context_for_test(run_id);
